relocate custom tags
